requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
coverage==4.5.4
Django==2.2.9
django-crispy-forms==1.7.2
djangorestframework==3.9.4
djangorestframework-datatables==0.5.0
factory-boy==2.12.0
Faker==2.0.2
idna==2.8
importlib-metadata==0.22
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
mock==3.0.5
more-itertools==7.2.0
mysqlclient==1.4.2
packaging==19.1
pluggy==0.13.0
py==1.8.0
pylint==2.3.1
pyparsing==2.4.2
pytest==5.1.2
pytest-cov==2.7.1
pytest-django==3.5.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
